Support subframe loading from SXG in subresource Web Bundles
To support subframe loading from SXG in subresource Web Bundles, this CL
changes BeginNavigation IPC to pass an URLLoaderClientEndpoints.
While navigating a subframe to a SXG content inside a Web Bundle which
was loaded using a link tag with rel=webbundle,
LinkWebBundle::MaybeCreateURLLoaderClientEndpoints() is called to create
an URLLoaderClientEndpoints which is used to load the SXG from the Web
Bundle. The URLLoaderClientEndpoints is passed through the
BeginNavigation IPC to the NavigationURLLoaderImpl.

Here are the calling relationships of the key methods:
  FrameLoader::StartNavigation
   ResourceFetcher::MaybeGetSubresourceWebBundleURLLoaderClientEndpoints
    LinkWebBundle::MaybeCreateURLLoaderClientEndpoints
  RenderFrameHostImpl::BeginNavigation
   Navigator::OnBeginNavigation
    NavigationRequest::CreateRendererInitiated
  NavigationRequest::OnStartChecksComplete
   NavigationURLLoader::Create
    NavigationURLLoaderImpl::NavigationURLLoaderImpl

When the URLLoaderClientEndpoints
|subresource_web_bundle_url_loader_client_endpoints_| is available, the
NavigationURLLoaderImpl will call
SubresourceWebBundleNavigationURLLoaderFactory::Create() to create a
SingleRequestURLLoaderFactory with the endpoint to load the resource
from the Web Bundle.

Currently we allow loading subframes from subresource Web Bundles only
when the resource is a signed exchange and the resource URL in the web
bundle and the inner URL of the signed exchange are the same.
A ResponseCheckingURLLoaderClient created by
SubresourceWebBundleNavigationURLLoaderFactory::Create() checks that the
resource is a signed exchange. And a SignedExchangeLoader checks that
the resource URL in the web bundle and the inner URL of the signed
exchange are the same.

Bug: 1082020
Change-Id: Id6a7d9adb6a056c458da546ed12ce9adc3180f4e