[WPT] Check worker.onerror message
This CL checks that worker.onerror event's `message` field
contains the original exception's message, "TypeError", etc.

This is a regression test for crbug/590219, where Chromium
set the message to just "Uncaught ".

This CL also refactors `exception-in-onerror.js` and renames it
to `throw.js`, to share it from multiple test files.

Bug: 590219
Change-Id: I7184eba7a9a6a7f08b1ef6eed95bc49adb91fcdd