Fix incorrect delay of DelayNode
WrapPositionVector was not wrapping the position around when it equaled
the end of the buffer.  This caused an extra zero value to be output
in this case.

Examination of WrapIndexVector also indicates a related logic error
because there's no >= intrinsic, so we used < but got the logic wrong.

The NEON version appears to be correct, so just added additional
comments to match the SSE2 version.

Added one test to make sure the delay is correct (based on the bug
report).

Bug: 1123023
Test: the-delaynode-interface/delay-test.html
Change-Id: Ibcb09573c53d62e926a332307d07601110b91aa3